Started working on the ability to start an effect when approaching the notch.
Some effects, such as Glow and Tesla, will want to start applying effects to the notch prior to touching the notch. This requires a larger tracking area which is currently attached to the NotchView with a negative inset (that seems to cause tracking problems so the contentView would be a better place). The mouse event methods on NotchEffect now have an underNotch convenience.