Fix cylon cursor blocking on multi-display setups
The inverted global coordinate space uses the top left of the _main_ screen as the origin. The regular screen coordinate space uses the bottom left of the _main_ screen as the origin. In order to translate any arbitrary screen coordinate from any possible screen with a vertical flip transform, we must know the height of the main screen.